+++ CHANGELOG.md
@@ -8,7 +8,7 @@ since `1.7.0`.
 
 ## [Unreleased]
 ### Added
-- Added the `TCOD_sys_get_error` and `TCOD_sys_set_error` functions.
+- Added the `TCOD_get_error` and `TCOD_set_error` functions.
 
 ### Changed
 - `TCOD_console_init_root` and `TCOD_console_set_custom_font` can now return